algorithm/legacy/pms/requirements.txt
joblib==1.3.2
numba==0.54.1
numpy==1.20.3
pandas==2.0.1
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.7.1
